changer le type de la colonne taille pour gérer la taille > 2Go

+++ migrations/20260209093232_ajout_taille_fichiers.js
@@ -4,7 +4,7 @@
  */
 export async function up(knex) {
     await knex.schema.alterTable('fichier', (table) => {
-        table.integer('taille')
+        table.bigint('taille')
     })
 
     await knex.raw(`UPDATE fichier SET taille = length(contenu)`)
